Extract helpers (#21)
